requirements.txt
a2a-sdk==0.3.25
flask==3.1.1
python-dotenv==1.2.2
